Model Update and Punch
In this build I used the baby model to find out how to properly rig
models. The baby model work with mecanim default animations and does not
stretch oddly when animations transpire.

I also added the punch mechanic to the baby model to experiment with it.
The baby can now punch blocks which will destroy them on impact. During
the punch animation colliders initiate around the arms that allow the
baby to destroy objects.

The Script I coded for this interaction is titled PunchButton_JG.