CharacterSwitch
I placed the script on the main camera because later what I am working
on for this script will likely eventually require the CameraFollow
script. Once you switch characters you need to tell the camera to lock
on to the new character.  Don't press the button twice (ie press "2"
while you are the teenager) because currently the inactive character is
not programmed to follow the position of the active character. I got
this working but there is oddities with Unity not disabling the
colliders (or rigidbody?) for the model for SetActive. I have solutions
that I will be working on.

Lastly there is a box collider on the character to prevent the inactive
character from falling through the world. Once I fix the issues we
should not needs this.

Controls - 1. Baby 2. Teen 3. Old (not yet implemented)