Character Switch Latest Update
Fixed the issues I had with the character switch. Greatly expanded the
code so everything can be set in inspector. Added comments to explain
how CharacterSwitch works. Made changes to Camerafollow so we can set
the models in inspector rather than depending on specific model names
(because they will change down the road).