Seed working in ProtoLevel3_NBMc_New
Seed is currently working in ProtoLevel3_NBMc_New.  With the update to
the EndLevelCube Prefab you can only end the level now if you have
picked up the seed.  In theory this means that it should work this way
in every level since the Player has a hasSeed boolean that starts the
level unchecked (false).  When the player picks up the seed this bool is
flip to true (Checked) allowing the player to move on to the end level
panel and thus the next scene.  Next up is fixing the panel so that it
give the option of taking the player to the next scene or restarting the
current scene and also giving the player the time it took them to
complete the level.