Seed/UITimer/PausePanels/UIButtons all working!
The Seed, UI Timer, EndPanel, Fail Panel and UI Buttons are all working
in NBMcShowtimeLevel.  Next step is to implement the prefabs into the
build scenes.

S/N to set the next level button you just have to put the scene you want
to go to in the editor.  Currently this would have to be done for every
scene but I will find a more elegant solution!