Fully Functioning Seed Scripts
Scripts:
SeedPickup_JG
Soil
PlayerInventory_JG

These scripts allow players to pickup a tree seed or a vine seed. Once a
seed is obtained, the seed is disabled and  the player inventory boolean
values are updated. When the player stands on  soil and if they have a
seed they can plant a platforming tree or a vine that players can walk
across. Whenever a seed is planted it disappears from inventory and the
original seeds are enabled.  Under Scenes > Semester2 > JoelGarcia
folders, is a scene titled SeedExperiment1, illustrates the mechanics
functioning.

With a tree seed press H to plant it. With a vine seed press G to plant
it.