WaterMeter GUI Functioning
Water GUI functions when the player gains or loses water the meter
reflects the amount of water gained or lost.  Also but in testing
controls into the access so water can be incremented/decremented with
the "p" and "o" keys respectively.

Also cleaned up the WaterManager_JG Script to have variables that are
adjustable within the inspector for min and max water amounts.